Add a scroll effect to the first step, adds an option to complete the tour on close and adds a callback to when the tour is closed before being completed.
There was no animation for the first step which resulted in popovers appearing below the fold, or above it in the event where the user reloaded a page and was at the bottom, not to be visible.

I also added the ability to end the tour completed if the users closes the popover.  There was no way for users that aren't interested in completing the tour not to go through all of the steps.  This solves that.  I didn't want to change the API too much and adding a second option that could have conflicted with `nextOnClose` wasn't ideal so I've chosen to make it complete the tour if `nextOnClose` is set to 'close'.  Setting it to true or any non `false` value will continue to work as it did before.

Finally, I added a callback that's trigger when the uses manually ends the tour.  The callback receives the last viewed step as it's first argument.